+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Provide an easy access link to the message centre in the `message list` command.
 - Show the submissions configuration data when rendering a study.
+- Provide the ability to create studies with eligibility requirements.
 
 ## 0.0.27